reddit_data_helper/backend_processing/requirements.txt
flask
flask-cors
langchain
google-generativeai
langchain-google-genai
python-dotenv
requests